requirements.txt
Django~=5.2.5
djangorestframework~=3.16.1
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0

django-filter~=25.1
requests~=2.32.5